separated a general component to its own place
